+++ packages/components/src/components/combobox/combobox.component.ts
@@ -22,8 +22,9 @@ import type SynOptGroup from '../optgroup/optgroup.js';
 import styles from './combobox.styles.js';
 import customStyles from './combobox.custom.styles.js';
 import {
-  createOptionFromDifferentTypes,
-  filterOnlyOptgroups, getAllOptions, getAssignedElementsForSlot, normalizeString,
+  checkValueBelongsToOption,
+  createOptionFromDifferentTypes, filterOnlyOptgroups, getAllOptions, getAssignedElementsForSlot,
+  getValueFromOption, normalizeString,
 } from './utils.js';
 import { scrollIntoView } from '../../internal/scroll.js';
 import { type OptionRenderer, defaultOptionRenderer } from './option-renderer.js';
@@ -77,6 +78,7 @@ import { enableDefaultSettings } from '../../utilities/defaultSettings/decorator
  * @csspart expand-icon - The container that wraps the expand icon.
  * @csspart popup - The popup's exported `popup` part.
  * Use this to target the tooltip's popup container.
+ * @csspart no-results - The container that wraps the "no results" message.
  *
  * @animation combobox.show - The animation to use when showing the combobox.
  * @animation combobox.hide - The animation to use when hiding the combobox.
@@ -106,8 +108,8 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
 
   private closeWatcher: CloseWatcher | null;
 
-  /** The last value of a syn-option, that was selected by click or via keyboard navigation */
-  private lastOptionValue = '';
+  /** The last syn-option, that was selected by click or via keyboard navigation */
+  private lastOption: SynOption | undefined;
 
   private isOptionRendererTriggered = false;
 
@@ -126,6 +128,8 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
   @query('slot:not([name])') private defaultSlot: HTMLSlotElement;
 
   @state() private hasFocus = false;
+
+  @state() private isUserInput = false;
 
   @state() displayLabel = '';
 
@@ -194,6 +198,12 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
 
   /** The combobox's required attribute. */
   @property({ reflect: true, type: Boolean }) required = false;
+
+  /**
+   * When set to `true`, restricts the combobox to only allow selection from the available options.
+   * Users will not be able to enter custom values that are not present in the list.
+   */
+  @property({ reflect: true, type: Boolean }) restricted = false;
 
   /**
    * A function that customizes the rendered option. The first argument is the option, the second
@@ -365,10 +375,11 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
 
       // Update the value based on the current selection and close it
       if (currentOption) {
-        const oldValue = this.lastOptionValue;
-        this.setSelectedOption(currentOption);
+        const oldValue = this.lastOption ? getValueFromOption(this.lastOption) : undefined;
+        this.updateSelectedOptionsCacheAndValue(currentOption);
 
         if (this.value !== oldValue) {
+          this.setSelectedOptionToSelected();
           // Emit after updating
           this.updateComplete.then(() => {
             this.emit('syn-input');
@@ -457,9 +468,10 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
     if (this.value !== '') {
       this.value = '';
       this.displayInput.value = '';
-      this.lastOptionValue = '';
-      this.setSelectedOption(undefined);
+      this.lastOption = undefined;
+      this.updateSelectedOptionsCacheAndValue(undefined);
       this.displayInput.focus({ preventScroll: true });
+      this.setSelectedOptionToSelected();
 
       // Emit after update
       // eslint-disable-next-line @typescript-eslint/no-floating-promises
@@ -482,14 +494,15 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
   private handleOptionClick(event: MouseEvent) {
     const target = event.target as HTMLElement;
     const option = target.closest('syn-option');
-    const oldValue = this.lastOptionValue;
+    const oldValue = this.lastOption ? getValueFromOption(this.lastOption) : undefined;
     if (option && !option.disabled) {
-      this.setSelectedOption(option);
+      this.updateSelectedOptionsCacheAndValue(option);
 
       // Set focus after updating so the value is announced by screen readers
       this.updateComplete.then(() => this.displayInput.focus({ preventScroll: true }));
 
       if (this.value !== oldValue) {
+        this.setSelectedOptionToSelected();
         // Emit after updating
         this.updateComplete.then(() => {
           this.emit('syn-input');
@@ -565,20 +578,20 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
   /**
    * Updates the selected options cache, the current value, and the display value
    */
-  private setSelectedOption(option: SynOption | undefined) {
+  // eslint-disable-next-line complexity
+  private updateSelectedOptionsCacheAndValue(option: SynOption | undefined) {
     this.selectedOption = option;
 
-    // Check if the selected option has a value,
-    // if not take the text content of the option otherwise the input text
     let optionValue;
-    if (this.selectedOption?.value) {
-      optionValue = String(this.selectedOption.value);
-    } else {
-      optionValue = this.selectedOption?.getTextLabel();
-    }
 
     if (option) {
-      this.lastOptionValue = optionValue || '';
+      this.lastOption = option;
+      optionValue = String(getValueFromOption(option));
+    } else if (this.restricted && !this.isValidValue() && this.value !== '' && !this.isUserInput) {
+      // if an invalid value was set via property binding for `restricted`comboboxes,
+      // reset to last valid value
+      this.resetToLastValidValue();
+      return;
     }
 
     // Update the value
@@ -590,6 +603,18 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
       this.displayLabel = this.selectedOption?.getTextLabel() ?? this.displayInput.value;
       this.formControlController.updateValidity();
     });
+  }
+
+  private setSelectedOptionToSelected() {
+    const slottedOptions = this.getSlottedOptions();
+    slottedOptions.forEach((opt) => {
+      // eslint-disable-next-line no-param-reassign
+      opt.selected = false;
+    });
+
+    if (this.selectedOption) {
+      this.selectedOption.selected = true;
+    }
   }
 
   private handleInvalid(event: Event) {
@@ -624,8 +649,8 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
   @watch('open', { waitUntilFirstUpdate: true })
   async handleOpenChange() {
     if (this.open && !this.disabled) {
-      if (this.numberFilteredOptions === 0) {
-        // Don't open the listbox if there are no options
+      if (this.numberFilteredOptions === 0 && !this.restricted) {
+        // Don't open the listbox if there are no options and it is not restricted
         this.open = false;
         this.emit('syn-error');
         return;
@@ -774,27 +799,92 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
 
   private async handleInput() {
     const inputValue = this.displayInput.value;
+    const cachedLastOption = this.lastOption;
+    this.isUserInput = true;
+
     this.value = inputValue;
     await this.updateComplete;
-    this.open = this.numberFilteredOptions > 0;
-    this.setSelectedOption(undefined);
-
+    this.isUserInput = false;
+    this.lastOption = cachedLastOption;
+    this.open = this.restricted || this.numberFilteredOptions > 0;
+    this.selectedOption = undefined;
     this.formControlController.updateValidity();
     this.emit('syn-input');
+  }
+
+  /**
+   * Checks if the current value is available in the options list.
+   * This is used to determine if the value is valid when the combobox is restricted.
+   *
+   * @returns `true` if the current value is available in the options list,
+   * otherwise `false`.
+   */
+  private isValidValue(): boolean {
+    const isValid = this.cachedOptions.some(
+      option => getValueFromOption(option) === this.value,
+    );
+    return isValid;
+  }
+
+  private getOptionFromValue(): SynOption | undefined {
+    return this.cachedOptions.find(option => checkValueBelongsToOption(this.value, option));
+  }
+
+  /**
+   * Resets the value to the last valid value or to an empty string.
+   */
+  private resetToLastValidValue() {
+    let label = '';
+    let value = '';
+
+    if (this.lastOption) {
+      value = String(getValueFromOption(this.lastOption));
+      label = this.lastOption.getTextLabel();
+    }
+
+    // Wait for the popup close animation to be finished before updating the value.
+    // This is to prevent flickering of the listbox, as the value is potentially reset to empty
+    // string and the whole options would be shown again.
+    const popupAnimations = this.popup?.popup?.getAnimations?.() ?? [];
+    const waitForAnimations = popupAnimations.length
+      ? Promise.all(
+        popupAnimations.map(animation => (animation.playState === 'finished'
+          ? Promise.resolve()
+          : new Promise<void>(resolve => { animation.addEventListener('finish', () => resolve(), { once: true }); }))),
+      )
+      : Promise.resolve();
+
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    waitForAnimations.then(() => {
+      this.value = value;
+      this.displayInput.value = label;
+      this.formControlController.updateValidity();
+    });
   }
 
   private handleChange() {
     // Only update the value and emit the event, if the change event occurred by
     // the user typing something in and removing focus of the combobox
-    if (!this.selectedOption) {
-      this.value = this.displayInput.value;
-      // Update validity
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-      this.updateComplete.then(() => {
-        this.formControlController.updateValidity();
-      });
-      this.emit('syn-change');
+    if (this.selectedOption) {
+      return;
     }
+
+    // If the value is not valid, we need to reset the value to the last valid value
+    if (this.restricted && !this.isValidValue() && this.value !== '') {
+      this.resetToLastValidValue();
+      return;
+    }
+
+    // Otherwise, update value from input and emit change
+    this.value = this.displayInput.value;
+    this.lastOption = this.getOptionFromValue();
+    this.selectedOption = this.getOptionFromValue();
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    this.updateComplete.then(() => {
+      this.formControlController.updateValidity();
+    });
+    this.setSelectedOptionToSelected();
+    this.emit('syn-change');
   }
 
   private getSlottedOptions() {
@@ -826,14 +916,13 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
   private updateSelectedOptionFromValue(): void {
     // check if the value has a corresponding option via value or text content
     // for empty values use the text content, as then the values of the option are not set
-    const option = this.cachedOptions
-      .find(o => (o.value !== '' && o.value === this.value) || (o.getTextLabel() === this.value));
+    const option = this.getOptionFromValue();
 
     if (!option) {
       this.displayInput.value = this.value;
     }
 
-    this.setSelectedOption(option);
+    this.updateSelectedOptionsCacheAndValue(option);
     this.createComboboxOptionsFromQuery(this.value);
   }
 
@@ -1005,6 +1094,14 @@ export default class SynCombobox extends SynergyElement implements SynergyFormCo
               @mouseup=${this.handleOptionClick}
             >
               <div class="listbox__options" part="filtered-listbox">
+                ${this.numberFilteredOptions === 0
+                  ? html`<span
+                      class="listbox__no-results"
+                      aria-hidden="true"
+                      part="no-results"
+                      >${this.localize.term('noResults')}</span
+                    >`
+                  : ''}
                 <slot @slotchange=${this.handleDefaultSlotChange}></slot>      
               </div>
             </div>