+++ packages/components/src/components/combobox/combobox.custom.styles.ts
@@ -11,5 +11,25 @@ export default css`
     display: none;
   }
 
+  .listbox__no-results {
+    align-items: center;
+    color: var(--syn-color-neutral-950);
+    display: flex;
+    font: var(--syn-body-medium-regular);
+    padding: var(--syn-spacing-small) var(--syn-spacing-medium) var(--syn-spacing-small) 52px;
+  }
+
+  .combobox--small .listbox__no-results {
+    font-size: var(--syn-input-font-size-small);
+    min-height: var(--syn-input-height-small);
+    padding: 0 var(--syn-spacing-small) 0 40px;
+  }
+
+  .combobox--large .listbox__no-results {
+    font-size: var(--syn-input-font-size-large);
+    min-height: var(--syn-input-height-large);
+    padding: 0 var(--syn-spacing-large) 0 68px;
+  }
+
   ${sharedOptionSize}
 `;